Add Godot project
Introduces the Godot project "GridMath" with two classes mCurve and
NestedGrid. The mCurve class represents a mathematical curve like the
Curve class from the python section - the class name Curve is being used
in Godot. The NestedGrid class is another variation of the 2D grid
class. Values are stored as floats in a nested Array.
These classes can be used to compare the performance of the python
implementations with the Godot implementations.