Add new class: FlatDictGrid
Introduces a new class FlatDictGrid. Just as the DictGrid it represents
a two-dimensional grid structure with a dictionary. Instead of using
tuples as keys, FlatDictGrid maps cell coordinates with the formula
x + y * width to their values. This new class allows to compare the
performance between creation of a structure and a calculation to access
values in a dictionary.