Add new Godot class: NestedGridRect
Introduces NestedGridRect to Godot project. It uses nested Arrays to
represents the two-dimensional grid structure. It makes uses of the
Rect2i class of Godot to store its position and size. The Rect2i class
has an intersection method to calculate the overlapping area of two
Rect2i instances. This should make the code more readable and a bit
faster since the intersection method is written in C++.