Add new Godot class: FlatGridRect
Introduces FlatGridRect to Godot project. It uses a 1D Array to
represents the two-dimensional grid structure. It makes uses of the
Rect2i class of Godot to store its position and size. The Rect2i class
has an intersection method to calculate the overlapping area of two
Rect2i instances. This class was created because it should be the
fastest Godot class yet, with NestedGrid being slower than
NestedGridRect and FlatGrid being faster than NestedGrid. But somehow
performance_test showed that the combination of the two features is
slower than the simple FlatGridRect.