Add C++ solution
Introduces the C++ Solution "GridMath" with two classes Curve and
FlatGrid. As in the python and Godot project the Curve class represents
a mathematical curve. The FlatGrid class is variation of the grid
class written in C++. Values are stored as floats in an array. The
solution also contains a Vector2i struct which is used to store
coordinates inside to integers. Together these classes can be used to
compare the performance to the python and Godot implementations.