Updated Performance Test
Updated all perfomance tests to compare the efficiency of the grids. To
test perfomance two instances of each class are created. A small grid
with size 9x9 and a larger 600x600 grid. The method add_grid_at_pos
is called for every cell of the large grid, adding the small grid
centered at that position. The test repeats 30 times for each class.