Add new class DictGrid
Introduces a new class DictGrid to Godot. Just as the FlatDictGrid it
represents a two-dimensional grid structure with a dictionary. Instead
of using Integers as keys, DictGrid maps cell coordinates with a
Vector2i to their values. This new class allows to compare the
performance between creation of a Vector2i and a calculation to access
values in a dictionary.